requirements.txt
ansicolors==1.1.8
ansiwrap==0.8.4
black==24.4.2
click==8.1.7
console-menu==0.8.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psutil==5.9.8
six==1.16.0
textwrap3==0.9.2
tomli==2.0.1
typing_extensions==4.12.2
